scraper_project/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
openpyxl==3.1.2
schedule==1.2.0
python-dotenv==1.0.0
lxml==4.9.3
fake-useragent==1.4.0
webdriver-manager==4.0.1
numpy==1.24.3
datetime
email-validator==2.1.0
smtplib
logging
json
re
time 